fix: #11392 Time Buttons Should be Scroll not the Whole Right Side
